add batch extender
