+++ sequencing/sequencer.go
@@ -3,7 +3,6 @@ package sequencing
 import (
 	"bytes"
 	"context"
-	"crypto/sha256"
 	"encoding/hex"
 	"errors"
 	"fmt"
@@ -38,7 +37,7 @@ type BatchQueue struct {
 	mu    sync.Mutex
 }
 
-// NewBatchQueue creates a new TransactionQueue
+// NewBatchQueue creates a new BatchQueue
 func NewBatchQueue() *BatchQueue {
 	return &BatchQueue{
 		queue: make([]sequencing.Batch, 0),
@@ -134,134 +133,6 @@ func (bq *BatchQueue) LoadFromDB(db *badger.DB) error {
 	return err
 }
 
-// TransactionQueue is a queue of transactions
-type TransactionQueue struct {
-	queue []sequencing.Tx
-	mu    sync.Mutex
-}
-
-// NewTransactionQueue creates a new TransactionQueue
-func NewTransactionQueue() *TransactionQueue {
-	return &TransactionQueue{
-		queue: make([]sequencing.Tx, 0),
-	}
-}
-
-// GetTransactionHash to get hash from transaction bytes using SHA-256
-func GetTransactionHash(txBytes []byte) string {
-	hashBytes := sha256.Sum256(txBytes)
-	return hex.EncodeToString(hashBytes[:])
-}
-
-// AddTransaction adds a new transaction to the queue
-func (tq *TransactionQueue) AddTransaction(tx sequencing.Tx, db *badger.DB) error {
-	tq.mu.Lock()
-	tq.queue = append(tq.queue, tx)
-	tq.mu.Unlock()
-
-	// Store transaction in BadgerDB
-	err := db.Update(func(txn *badger.Txn) error {
-		return txn.Set([]byte(GetTransactionHash(tx)), tx)
-	})
-	return err
-}
-
-// GetNextBatch extracts a batch of transactions from the queue
-func (tq *TransactionQueue) GetNextBatch(max uint64, db *badger.DB) sequencing.Batch {
-	tq.mu.Lock()
-	defer tq.mu.Unlock()
-
-	var batch [][]byte
-	batchSize := len(tq.queue)
-	if batchSize == 0 {
-		return sequencing.Batch{Transactions: nil}
-	}
-	for {
-		batch = tq.queue[:batchSize]
-		blobSize := totalBytes(batch)
-		if uint64(blobSize) <= max {
-			break
-		}
-		batchSize = batchSize - 1
-	}
-
-	// Retrieve transactions from BadgerDB and remove processed ones
-	for _, tx := range batch {
-		txHash := GetTransactionHash(tx)
-		err := db.Update(func(txn *badger.Txn) error {
-			// Get and then delete the transaction from BadgerDB
-			_, err := txn.Get([]byte(txHash))
-			if err != nil {
-				return err
-			}
-			return txn.Delete([]byte(txHash)) // Remove processed transaction
-		})
-		if err != nil {
-			return sequencing.Batch{Transactions: nil} // Return empty batch if any transaction retrieval fails
-		}
-	}
-	tq.queue = tq.queue[batchSize:]
-	return sequencing.Batch{Transactions: batch}
-}
-
-// LoadFromDB reloads all transactions from BadgerDB into the in-memory queue after a crash.
-func (tq *TransactionQueue) LoadFromDB(db *badger.DB) error {
-	tq.mu.Lock()
-	defer tq.mu.Unlock()
-
-	// Start a read-only transaction
-	err := db.View(func(txn *badger.Txn) error {
-		// Create an iterator to go through all transactions stored in BadgerDB
-		it := txn.NewIterator(badger.DefaultIteratorOptions)
-		defer it.Close() // Ensure that the iterator is properly closed
-
-		// Iterate through all items in the database
-		for it.Rewind(); it.Valid(); it.Next() {
-			item := it.Item()
-			err := item.Value(func(val []byte) error {
-				// Load each transaction from DB and add to the in-memory queue
-				tq.queue = append(tq.queue, val)
-				return nil
-			})
-			if err != nil {
-				return err
-			}
-		}
-		return nil
-	})
-
-	return err
-}
-
-// AddBatchBackToQueue re-adds the batch to the transaction queue (and BadgerDB) after a failure.
-func (tq *TransactionQueue) AddBatchBackToQueue(batch sequencing.Batch, db *badger.DB) error {
-	tq.mu.Lock()
-	defer tq.mu.Unlock()
-
-	// Add the batch back to the in-memory transaction queue
-	tq.queue = append(tq.queue, batch.Transactions...)
-
-	// Optionally, persist the batch back to BadgerDB
-	for _, tx := range batch.Transactions {
-		err := db.Update(func(txn *badger.Txn) error {
-			return txn.Set([]byte(GetTransactionHash(tx)), tx) // Store transaction back in DB
-		})
-		if err != nil {
-			return fmt.Errorf("failed to revert transaction to DB: %w", err)
-		}
-	}
-
-	return nil
-}
-
-func totalBytes(data [][]byte) int {
-	total := 0
-	for _, sub := range data {
-		total += len(sub)
-	}
-	return total
-}
-
 // Sequencer implements go-sequencing interface
 type Sequencer struct {
 	dalc      *da.DAClient
@@ -286,7 +157,7 @@ type Sequencer struct {
 }
 
 // NewSequencer ...
-func NewSequencer(daAddress, daAuthToken string, daNamespace []byte, rollupId []byte, batchTime time.Duration, metrics *Metrics, dbPath string) (*Sequencer, error) {
+func NewSequencer(daAddress, daAuthToken string, daNamespace []byte, rollupId []byte, batchTime time.Duration, metrics *Metrics, dbPath string, extender BatchExtender) (*Sequencer, error) {
 	ctx := context.Background()
 	dac, err := proxyda.NewClient(daAddress, daAuthToken)
 	if err != nil {
@@ -316,7 +187,7 @@ func NewSequencer(daAddress, daAuthToken string, daNamespace []byte, rollupId []
 		ctx:         ctx,
 		maxSize:     maxBlobSize,
 		rollupId:    rollupId,
-		tq:          NewTransactionQueue(),
+		tq:          NewTransactionQueue(extender),
 		bq:          NewBatchQueue(),
 		seenBatches: make(map[string]struct{}),
 		db:          db,
